Fix failing Docker image builds. (#28)
* Switch Alpine images to use Python 3.X.

They appear to have dropped a number of Python 2.X packages (probably
due to Python 2.X being EOL as-of 2020-04-01). This switches to using
Python 3.X with the appropriate changes to the Python packages being
pulled in.

* Change what MySQL Python module is used on Fedora 31.

MySQL-python is no longer a thing in Fedora 31. Luckily they do have the
other Python MySQL client that we are able to use, so switch to using
that instead.