Add package builder images for Debian. (#31)
* Add package builder images for Debian.

This adds dockerfiles for images to be used for building packages for
Debian. The resultant image have everything needed for building Netdata
packages except for the netdata sources, which are intended to be
mounted as a volume on `/host/netdata` from the host system.

Once started with the sources mounted, one can simply run `/build.sh` to
build the packages, which will be copied into the root of the sources
once built.

* Add variables to control default changelog contents.

* Drop superfluous ENV lines.

Also, use the current date instead of a fixed one when no date is
specified.

* Move the expected source location for consistence with other images.

* Re-add ENV lines. They actually are needed.

* Update package builder script to clean up after itself.

This ensures that the sources remain clean (other than the generated
package files) each time the package builder is run.

* Explicitly use the nointeractive frontend for dpkg.

This adds an environment variable during the build process that will
prevent dpkg from prompting the user for type of input.

* Remove Debian Bookworm dockerfile.

It's two releases out so there's no way to test it.