Better isolate the sources from the build process.
This changes how the source and artifacts directories are handled during
builds for Debian packages in the following ways:

* The actual build is run in a copy of the sources that's local to the
  container. This ensures that the build leaves nothing it shouldn't in
  the user's source directory.
* The artifacts directory and it's contents are modified to have the
  same ownership as the original source directory.

Both of these are important on Linux systems because the scripts inside
the container run as UID 0, which means that the files and directories
produced in the source directory by the build are owned by root, and
threfore it may not be possible for a regular user to remove them
without explicitly elevating to EUID 0.

This is a non-issue on Mac and Windows systems because of how they
handle Docker.