Consolidate the layers in the base image.
This collapses the base image used for building our Docker images into a
single layer on top of the base Alpine image. None of the things we're
doing are realistically cachable (at least, not safely), so there's no
advantage to having multiple layers for this.

This should cut down on final image sizes and improve pull times.