Add Clang to build environments. (#333)
We’re looking at possibly switching from GCC to Clang to take advantage
of faster builds (and possibly also ThinLTO) in CI, so we need it
available in all of our build environments. This does not enable usage
of it for any of our builds yet (that will be handled in the actual
agent repository, not here).